refactored bindings in mulitple files
