+++ src/bindings/bindings.hpp
@@ -1,0 +1,18 @@
+// SPDX-FileCopyrightText: 2024-2026 NeoFOAM authors
+// SPDX-License-Identifier: Unlicense
+
+#pragma once
+
+#include <nanobind/nanobind.h>
+
+namespace NeoFOAM::bindings
+{
+
+void registerRuntime(nanobind::module_& m);
+void registerFieldFactories(nanobind::module_& m);
+void registerFieldWriters(nanobind::module_& m);
+void registerPDESolver(nanobind::module_& m);
+void registerPressureVelocityCoupling(nanobind::module_& m);
+void registerUtility(nanobind::module_& m);
+
+} // namespace NeoFOAM::bindings
